+++ app/include_upgrade_mainnet.go
@@ -4,8 +4,6 @@ package app
 
 import (
 	"github.com/babylonlabs-io/babylon/v4/app/upgrades"
-	v1 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v1"
-	"github.com/babylonlabs-io/babylon/v4/app/upgrades/v1/mainnet"
 	"github.com/babylonlabs-io/babylon/v4/app/upgrades/v1_1"
 	v2 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v2"
 	v22 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v2_2"
@@ -31,13 +29,5 @@ func init() {
 		v22.Upgrade,
 		v2.CreateUpgrade(false, WhitelistedChannelsID),
 		v1_1.Upgrade,
-		v1.CreateUpgrade(v1.UpgradeDataString{
-			BtcStakingParamsStr:       mainnet.BtcStakingParamsStr,
-			FinalityParamStr:          mainnet.FinalityParamStr,
-			IncentiveParamStr:         mainnet.IncentiveParamStr,
-			CosmWasmParamStr:          mainnet.CosmWasmParamStr,
-			NewBtcHeadersStr:          mainnet.NewBtcHeadersStr,
-			TokensDistributionStr:     mainnet.TokensDistributionStr,
-			AllowedStakingTxHashesStr: mainnet.AllowedStakingTxHashesStr,
-		}, mainnet.ParamUpgrade)}
+	}
 }